move a few name table checks out of microsoft directory
* name_id_1
* name_id_2
* name_length_req
* typographic_family_name
